fix: float64 accumulator in CM conservation, load-once GAUL optimization
CM conservation assertion used float32 nansum, which would produce false
positives from summation order sensitivity when feature totals exceed 2^24.
Fix: accumulate in float64 before comparing.

GAUL area-majority script loaded the shapefile and built the STRtree 7×
(once per variable). The geometric computation is identical across all
variables — only the field lookup differs. Refactored to load once, join
once, extract per-field values from the shared polygon index map. Expected
Step 1 runtime reduction: ~2.3h → ~20min on server.

Also fixes risk register header count (192 → 195 resolved).

+++ src/datafactory_adapters/_conservation.py
@@ -32,9 +32,9 @@ def assert_cm_conservation(
         return
 
     for i in indices:
-        grid_total = float(np.nansum(flat_data_all[:, i]))
-        land_total = float(np.nansum(flat_data_land[:, i]))
-        excl_total = float(np.nansum(excluded_data[:, i]))
+        grid_total = float(np.nansum(flat_data_all[:, i], dtype=np.float64))
+        land_total = float(np.nansum(flat_data_land[:, i], dtype=np.float64))
+        excl_total = float(np.nansum(excluded_data[:, i], dtype=np.float64))
         recon = land_total + excl_total
 
         if not np.allclose(grid_total, recon, rtol=1e-6, atol=1e-4):
